option UI
옵션 버튼 누르면 옵션창이 열리고 옵션 나가기 버튼 누르면 옵션창이 닫히는 기능 구현
